refactor: update import paths for consistency in MCP and testing modules
