+++ tests/conftest.py
@@ -14,7 +14,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from src.aurite.lib.config.config_models import HostConfig
+from src.aurite.lib.models.config.components import HostConfig
 
 logging.basicConfig(level=logging.INFO, format="%(levelname)s - %(name)s - %(message)s")
 logger = logging.getLogger(__name__)
